Potential fix for code scanning alert no. 400: Multiplication result converted to larger type (#1048)

+++ src/generator_packed_spgemm_bcsc_bsparse_aarch64.c
@@ -1189,7 +1189,7 @@ void libxsmm_generator_packed_spgemm_bcsc_bsparse_aarch64_kloop_bfdot_sve(libxsm
                                                          i_gp_reg_mapping->gp_reg_c,
                                                          l_gp_reg_scratch,
                                                          i_gp_reg_mapping->gp_reg_c,
-                                                         ((long long)i_packed_width - i_packed_blocking * i_simd_packed_width) * i_micro_kernel_config->datatype_size_out );
+                                                         ((long long)i_packed_width - ((long long)i_packed_blocking * i_simd_packed_width)) * i_micro_kernel_config->datatype_size_out );
         }
       }
       libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code,
